DBProvisioningDemo/requirements.txt
click>=8.1
rich>=13.0
psycopg2-binary>=2.9
boto3>=1.34
kubernetes>=29.0
python-dotenv>=1.0
pydantic>=2.0
pydantic-settings>=2.0
